+++ tests/test_helpers.py
@@ -5,11 +5,7 @@ from docfx_yaml.extension import convert_cross_references
 from docfx_yaml.extension import search_cross_references
 from docfx_yaml.extension import format_code
 from docfx_yaml.extension import extract_product_name
-from docfx_yaml.extension import highlight_md_codeblocks
-from docfx_yaml.extension import prepend_markdown_header
-from docfx_yaml.extension import clean_image_links
 from docfx_yaml.extension import reformat_summary
-from docfx_yaml.extension import reformat_markdown_to_html
 
 import unittest
 from parameterized import parameterized
@@ -246,100 +242,6 @@ for i in range(10):
         self.assertEqual(short_name_want, short_product_name)
 
 
-    # Filenames to test markdown syntax highlight with.
-    test_markdown_filenames = [
-        [
-            "tests/markdown_syntax_highlight.md",
-            "tests/markdown_syntax_highlight_want.md"
-        ],
-        [
-            "tests/markdown_no_highlight.md",
-            "tests/markdown_no_highlight_want.md"
-        ],
-        [
-            "tests/markdown_mixed_highlight.md",
-            "tests/markdown_mixed_highlight_want.md"
-        ],
-    ]
-    @parameterized.expand(test_markdown_filenames)
-    def test_highlight_md_codeblocks(self, base_filename, want_filename):
-        # Test to ensure codeblocks in markdown files are correctly highlighted.
-
-        # Copy the base file we'll need to test.
-        with tempfile.NamedTemporaryFile(mode='r+', delete=False) as test_file:
-            with open(base_filename) as base_file:
-                test_file.write(base_file.read())
-                test_file.flush()
-
-            highlight_md_codeblocks(test_file.name)
-            test_file.seek(0)
-
-            with open(want_filename) as mdfile_want:
-                self.assertEqual(test_file.read(), mdfile_want.read())
-
-
-    # Filenames to test prepending Markdown title..
-    test_markdown_filenames = [
-        [
-            "tests/markdown_example_bad_header.md",
-            "tests/markdown_example_bad_header_want.md"
-        ],
-        [
-            "tests/markdown_example_h2.md",
-            "tests/markdown_example_h2_want.md"
-        ],
-        [
-            "tests/markdown_example_alternate_bad.md",
-            "tests/markdown_example_alternate_bad_want.md"
-        ],
-    ]
-    @parameterized.expand(test_markdown_filenames)
-    def test_prepend_markdown_header(self, base_filename, want_filename):
-        # Ensure markdown titles are correctly prepended.
-
-        # Copy the base file we'll need to test.
-        with tempfile.NamedTemporaryFile(mode='r+', delete=False) as test_file:
-            with open(base_filename) as base_file:
-                # Use same file name extraction as original code.
-                file_name = base_file.name.split("/")[-1].split(".")[0].capitalize()
-                test_file.write(base_file.read())
-                test_file.flush()
-                test_file.seek(0)
-
-            prepend_markdown_header(file_name, test_file)
-            test_file.seek(0)
-
-            with open(want_filename) as mdfile_want:
-                self.assertEqual(test_file.read(), mdfile_want.read())
-
-
-    # Filenames to test cleaning up markdown image links.
-    test_markdown_filenames = [
-        [
-            "tests/markdown_example_bad_image_links.md",
-            "tests/markdown_example_bad_image_links_want.md"
-        ],
-    ]
-    @parameterized.expand(test_markdown_filenames)
-    def test_clean_image_links(self, base_filename, want_filename):
-        # Ensure image links are well formed in markdown files.
-
-        # Copy the base file we'll need to test.
-        with tempfile.NamedTemporaryFile(mode='r+', delete=False) as test_file:
-            with open(base_filename) as base_file:
-                # Use same file name extraction as original code.
-                file_name = base_file.name.split("/")[-1].split(".")[0].capitalize()
-                test_file.write(base_file.read())
-                test_file.flush()
-                test_file.seek(0)
-
-            clean_image_links(test_file.name)
-            test_file.seek(0)
-
-            with open(want_filename) as mdfile_want:
-                self.assertEqual(test_file.read(), mdfile_want.read())
-
-
     test_reference_params = [
         [
             # If no reference keyword is found, check for None
@@ -433,34 +335,6 @@ For example:
     def test_reformat_summary(self, summary, summary_want):
         summary_got = reformat_summary(summary)
         self.assertEqual(summary_want, summary_got)
-
-
-    test_markdown_content = [
-        [
-            """The resource name or `None`
-
-if no Cloud KMS key was used, or the blob's resource has not been loaded from the server.
-
-For example:
-```
-    kms_key_name: ID
-```
-            """,
-            """The resource name or <code>None</code>
-
-if no Cloud KMS key was used, or the blob's resource has not been loaded from the server.
-
-For example:
-<pre>
-    kms_key_name: ID
-</pre>
-            """,
-        ],
-    ]
-    @parameterized.expand(test_markdown_content)
-    def test_reformat_markdown_to_html(self, content, content_want):
-        content_got = reformat_markdown_to_html(content)
-        self.assertEqual(content_want, content_got)
 
 
 if __name__ == '__main__':
