+++ tests/test_unit.py
@@ -8,9 +8,7 @@ from docfx_yaml.extension import _extract_docstring_info
 from docfx_yaml.extension import find_package_group
 from docfx_yaml.extension import pretty_package_name
 from docfx_yaml.extension import group_by_package
-from docfx_yaml.extension import extract_header_from_markdown
 from docfx_yaml.extension import _parse_docstring_summary
-from docfx_yaml.extension import parse_markdown_header
 
 import unittest
 from parameterized import parameterized
@@ -591,147 +589,6 @@ Raises:
         toc_yaml_got = group_by_package(toc_yaml)
 
         self.assertCountEqual(toc_yaml_got, toc_yaml_want)
-
-
-    def test_parse_markdown_header(self):
-        # Test for simple header_line.
-        header_line_want = "Test header"
-
-        header_line = "# Test header"
-        prev_line = ""
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-        # Test for invalid input.
-        header_line_want = ""
-
-        header_line = "#Test header"
-        prev_line = ""
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-        # Test for invalid input.
-        header_line_want = ""
-
-        header_line = "#  Test header"
-        prev_line = ""
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-        # Test for no header.
-        header_line_want = ""
-
-        header_line = "-->"
-        prev_line = "limitations under the License.\n"
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-
-    def test_parse_markdown_header_alternate(self):
-        # Test for simple alternate header.
-        header_line_want = "Test header"
-
-        header_line = "============\n"
-        prev_line = "Test header"
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-        # Test for no header.
-        header_line_want = ""
-
-        header_line = "============\n"
-        prev_line = ""
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-
-        # Test for shorter divider.
-        header_line_want = "Test header"
-
-        header_line = "======\n"
-        prev_line = "Test header"
-
-        header_line_got = parse_markdown_header(header_line, prev_line)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-
-    test_markdown_filenames = [
-        [
-            # Check the header for a normal markdown file.
-            "tests/markdown_example.md"
-        ],
-        [
-            # The header should be the same even with the license header.
-            "tests/markdown_example_header.md"
-        ],
-    ]
-    @parameterized.expand(test_markdown_filenames)
-    def test_extract_header_from_markdown(self, markdown_filename):
-        # Check the header for markdown files.
-        header_line_want = "Test header for a simple markdown file."
-
-        with open(markdown_filename, 'r') as mdfile:
-            header_line_got = extract_header_from_markdown(mdfile)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-
-    test_markdown_filenames = [
-        [
-            # Check the header for an alternate header style.
-            "tests/markdown_example_alternate.md"
-        ],
-        [
-            # The header should be the same even with the license header.
-            "tests/markdown_example_alternate_header.md"
-        ],
-        [
-            # Check the header for an alternate header style.
-            "tests/markdown_example_alternate_less.md"
-        ],
-    ]
-    @parameterized.expand(test_markdown_filenames)
-    def test_extract_header_from_markdown_alternate_header(self, markdown_filename):
-        # Check the header for different accepted styles.
-        header_line_want = "This is a simple alternate header"
-
-        with open(markdown_filename, 'r') as mdfile:
-            header_line_got = extract_header_from_markdown(mdfile)
-
-        self.assertEqual(header_line_got, header_line_want)
-
-
-    test_markdown_filenames = [
-        [
-            "tests/markdown_example_bad_header.md"
-        ],
-        [
-            "tests/markdown_example_h2.md"
-        ],
-        [
-            "tests/markdown_example_alternate_bad.md"
-        ],
-    ]
-    @parameterized.expand(test_markdown_filenames)
-    def test_extract_header_from_markdown_bad_headers(self, markdown_filename):
-        # Check that empty string is returned if no valid header is found.
-        with open(markdown_filename, 'r') as mdfile:
-            header_line_got = extract_header_from_markdown(mdfile)
-
-        self.assertFalse(header_line_got)
 
 
     def test_parse_docstring_summary(self):
