+++ tox.ini
@@ -22,7 +22,7 @@ commands =
 deps =
     {[testenv]deps}
 commands =
-    python3 -m unittest tests/test_unit.py tests/test_helpers.py
+    python3 -m unittest tests/test_unit.py tests/test_helpers.py tests/test_markdown.py
 
 [testenv:librarytest]
 deps =
